Apply changes from feedback
